lp: Add Model class to represent a generic model
In linear programming, before running the Simplex algorithm using
the tableau representation, we need to convert the contraints and
the cost function into the standard form.

The standard form needs to be converted to the canonical form,
introducting several artificial variables, and a second cost
function.

This second representation can be used to calculate (if exist) the
first basic feasible soltuion, and from there (after removing the
artificial variables) execute the Simplex algorithm.